LayoutTests/http/tests/xmlhttprequest/xmlhttprequest-unsafe-redirect-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8080/xmlhttprequest/resources/forbidden.txt. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
This tests that unsafe redirects won't be allowed when making an XMLHttpRequest.
Sync XHR started.
readyState change 1
readyState change 4
Error event.
Exception: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Async XHR started.
readyState change 1
readyState change 4
SUCCESS: Error handler was called with readyState 4
